+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Added:
 - Previously sent messages can be accessed per buffer in the text input with up / down arrows
 - Themes directory where users can add their own theme files
 - Nickname completions in text input with <kbd>Tab</kbd>
+- Broadcast nickname changes to relevant channels and queries.
+- Broadcast quit messages to relevant channels and queries.
 
 Changed: